[Block Library]Rename QueryLoop to PostTemplate and change Query label
