+++ packages/block-library/src/editor.scss
@@ -46,7 +46,7 @@
 @import "./text-columns/editor.scss";
 @import "./video/editor.scss";
 @import "./query-title/editor.scss";
-@import "./query-loop/editor.scss";
+@import "./post-template/editor.scss";
 @import "./query/editor.scss";
 @import "./query-pagination/editor.scss";
 @import "./query-pagination-numbers/editor.scss";
