+++ packages/block-library/src/post-template/editor.scss
@@ -1,5 +1,5 @@
 .editor-styles-wrapper {
-	ul.wp-block-query-loop {
+	ul.wp-block-post-template {
 		padding-left: 0;
 		margin-left: 0;
 		list-style: none;
