+++ packages/block-library/src/post-template/style.scss
@@ -1,4 +1,4 @@
-.wp-block-query-loop {
+.wp-block-post-template {
 	max-width: 100%;
 	list-style: none;
 	padding: 0;
