+++ packages/block-library/src/query/block.json
@@ -1,7 +1,7 @@
 {
 	"apiVersion": 2,
 	"name": "core/query",
-	"title": "Query",
+	"title": "Query Loop",
 	"category": "design",
 	"description": "Displays a list of posts as a result of a query.",
 	"textdomain": "default",
