+++ packages/block-library/src/style.scss
@@ -34,7 +34,7 @@
 @import "./post-title/style.scss";
 @import "./preformatted/style.scss";
 @import "./pullquote/style.scss";
-@import "./query-loop/style.scss";
+@import "./post-template/style.scss";
 @import "./query-pagination/style.scss";
 @import "./quote/style.scss";
 @import "./rss/style.scss";
