+++ packages/e2e-tests/fixtures/blocks/core__post-template.html
@@ -1,0 +1,1 @@
+<!-- wp:post-template /-->
