+++ packages/e2e-tests/fixtures/blocks/core__post-template.json
@@ -1,7 +1,7 @@
 [
 	{
 		"clientId": "_clientId_0",
-		"name": "core/query-loop",
+		"name": "core/post-template",
 		"isValid": true,
 		"attributes": {},
 		"innerBlocks": [],
