+++ packages/e2e-tests/fixtures/blocks/core__post-template.parsed.json
@@ -1,6 +1,6 @@
 [
 	{
-		"blockName": "core/query-loop",
+		"blockName": "core/post-template",
 		"attrs": {},
 		"innerBlocks": [],
 		"innerHTML": "",
