--- packages/e2e-tests/fixtures/blocks/core__query-loop.html
@@ -1,1 +1,0 @@
-<!-- wp:query-loop /-->
