--- packages/e2e-tests/fixtures/blocks/core__query-loop.serialized.html
@@ -1,1 +1,0 @@
-<!-- wp:query-loop /-->
