+++ packages/e2e-tests/plugins/query-block.php
@@ -17,9 +17,9 @@ register_block_pattern(
 		'title'      => __( 'Query Test 1', 'gutenberg' ),
 		'blockTypes' => array( 'core/query' ),
 		'content'    => '<!-- wp:query {"query":{"perPage":1,"pages":0,"offset":0,"postType":"post","categoryIds":[],"tagIds":[],"order":"desc","orderBy":"date","author":"","search":"","exclude":[],"sticky":"","inherit":true}} -->
-						<!-- wp:query-loop -->
+						<!-- wp:post-template -->
 						<!-- wp:post-title {"isLink":true} /-->
-						<!-- /wp:query-loop -->
+						<!-- /wp:post-template -->
 						<!-- /wp:query -->',
 	)
 );
@@ -29,10 +29,10 @@ register_block_pattern(
 		'title'      => __( 'Query Test 2', 'gutenberg' ),
 		'blockTypes' => array( 'core/query' ),
 		'content'    => '<!-- wp:query {"query":{"perPage":1,"pages":0,"offset":0,"postType":"post","categoryIds":[],"tagIds":[],"order":"desc","orderBy":"date","author":"","search":"","exclude":[],"sticky":"","inherit":true}} -->
-						<!-- wp:query-loop -->
+						<!-- wp:post-template -->
 						<!-- wp:post-title {"isLink":true} /-->
 						<!-- wp:post-date /-->
-						<!-- /wp:query-loop -->
+						<!-- /wp:post-template -->
 						<!-- /wp:query -->',
 	)
 );
